+++ sdk/communication/azure-communication-administration/CHANGELOG.md
@@ -7,8 +7,8 @@
 - Added support for Azure Active Directory Authentication.
 
 ### Breaking Changes
-- CommunicationIdentityClient and CommunicationIdentityAsyncClient is moved to a new package, `azure-communication-identity`.
-- PhoneNumbersClient and PhoneNumbersAsyncClient is moved to a new package, `azure-communication-phonenumbers`.
+- CommunicationIdentityClient and CommunicationIdentityAsyncClient is moved to a new package, `azure.communication.identity`.
+- PhoneNumbersClient and PhoneNumbersAsyncClient is moved to a new package, `azure.communication.phonenumbers`.
 
 ## 1.0.0-beta.3 (2020-11-16)
 ### Added